remove no use code

+++ app/src/main/java/com/puutaro/commandclick/activity_lib/InitManager.kt
@@ -266,7 +266,7 @@ class InitManager(
         })
             .show()
         alertDialog.getButton(DialogInterface.BUTTON_POSITIVE).setTextColor(
-            activity.getColor(android.R.color.black) as Int
+            activity.getColor(android.R.color.black)
         )
         alertDialog.getButton(DialogInterface.BUTTON_NEGATIVE).setTextColor(
             activity.getColor(android.R.color.black)
